+++ src/providers/Providers.tsx
@@ -4,6 +4,7 @@ import { ReactNode } from 'react';
 import { PrimeReactProvider } from './PrimeReactProvider';
 import { ThemeProvider } from './ThemeProvider';
 import { HeaderProvider } from './HeaderProvider';
+import { CasesProvider } from './CasesProvider';
 
 interface ProvidersProps {
   children: ReactNode;
@@ -13,7 +14,9 @@ export const Providers = ({ children }: ProvidersProps) => {
   return (
     <ThemeProvider>
       <HeaderProvider>
-        <PrimeReactProvider>{children}</PrimeReactProvider>
+        <CasesProvider>
+          <PrimeReactProvider>{children}</PrimeReactProvider>
+        </CasesProvider>
       </HeaderProvider>
     </ThemeProvider>
   );